tools: move zoom_in in its own module
and fix all the dead code warnings on --no-default-features
